\name{CombinePDFs}

\alias{CombinePDFs}

\title{Combine PDF Files}

\description{
This function combines PDF files into a new file.
}

\usage{
CombinePDFs(path, open.pdf = FALSE)
}

\arguments{
\item{path}{character; something.}
\item{open.pdf}{logical; something.}
}

\details{
Something
}

\author{J.C. Fisher}

\examples{\donttest{
output.pdf <- CombinePDFs("path", open.pdf = TRUE)
}}

\keyword{misc}
