\name{GetPath}

\alias{GetPath}

\title{Get Path to File or Directory}

\description{
Something
}

\usage{
GetPath(type, path = NULL, initial.dir = getwd())
}

\arguments{
\item{type}{character; something.}
\item{path}{character; something.}
\item{initial.dir}{character; something.}
}

\details{
Something
}

\author{J.C. Fisher}

\examples{
print("something")
}

\keyword{misc}
