\name{OpenGraphicsDevice}

\alias{OpenGraphicsDevice}

\title{Open Graphics Device}

\description{
Starts the graphics device driver for producing graphics.
}

\usage{
OpenGraphicsDevice(gr.type, file.name, figs.dir,
                   w = 8.5, h = 11, p = 12, png.res = 300)
}

\arguments{
\item{gr.type}{character; something.}
\item{file.name}{character; something.}
\item{figs.dir}{character; something.}
\item{w}{numeric; something.}
\item{h}{numeric; something.}
\item{p}{numeric; something.}
\item{png.res}{numeric; something.}
}

\details{
Something
}

\value{
Something
}

\author{Fisher, J.C.}

\seealso{
\code{\link{windows}}, \code{\link{pdf}}, \code{\link{png}},
\code{\link{postscript}}
}

\examples{
\donttest{OpenGraphicsDevice()}
}

\keyword{misc}
