\name{PlotTrendData}

\alias{PlotTrendData}

\title{Plot Trend Data}

\description{
This function draws plots on the desired device type.
}

\usage{
PlotTrendData(d, site.names, sdate = NA, edate = NA, initial.dir = getwd(),
              file.symbs = NULL, file.plots = NULL, figs.dir = NULL,
              gr.type = "pdf")
}

\arguments{
\item{d}{data.frame; something.}
\item{site.names}{character; something.}
\item{sdate}{character; something.}
\item{edate}{character; something.}
\item{initial.dir}{character; something.}
\item{file.symbs}{character; something.}
\item{file.plots}{character; something.}
\item{figs.dir}{character; something.}
\item{gr.type}{character; something.}
}

\details{
Something
}

\value{
Something
}

\author{Fisher, J.C., Davis, L.C.}

\seealso{
\code{\link{plot.window}}, \code{\link{axis.POSIXct}}, \code{\link{legend}}
}

\examples{
\donttest{PlotTrendData()}
}

\keyword{hplot}
