% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessWL.R
\name{ProcessWL}
\alias{ProcessWL}
\title{Process Water Levels}
\usage{
ProcessWL(water.levels, date.fmt = "\%Y-\%m-\%d \%H:\%M")
}
\arguments{
\item{water.levels}{data.frame.
See \sQuote{Details} section.}

\item{date.fmt}{character.
Date format used to convert character strings to Date class.}
}
\value{
Returns a data.frame object with the following components:
  \describe{
    \item{Site_id}{unique site identifier}
    \item{Date}{measurement date}
    \item{Var}{water-level elevation}
  }
}
\description{
This function processes water levels for data analysis.
}
\details{
Required columns in the \code{water.levels} data table include:
  \code{SITE_NO}, a unique site identifier;
  \code{LEV_DT}, the measurement date-time;
  \code{ALT_VA}, referenced land-surface elevation; and
  \code{LEV_VA}, depth below land surface to the water table.
  The \code{water.levels} data table is composed of character-class components.
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\seealso{
\code{\link{RunAnalysis}}
}
\keyword{methods}

