\name{RunStats}

\alias{RunStats}

\title{Open Graphics Device}

\description{
This function performs a statistical analysis on uncensored and censored data.
}

\usage{
RunStats(d, site.names, is.censored = FALSE, initial.dir = getwd(),
         file.stats = NULL, file.out = NULL, figs.dir = NULL,
         avg.time = "year", gr.type = "pdf")
}

\arguments{
\item{d}{data.frame; something.}
\item{site.names}{character; something.}
\item{is.censored}{logical; something.}
\item{initial.dir}{character; something.}
\item{file.stats}{character; something.}
\item{file.out}{character; something.}
\item{figs.dir}{character; something.}
\item{avg.time}{character; something.}
\item{gr.type}{character; something.}
}

\details{
Something
}

\value{
Something
}

\author{Fisher, J.C., Davis, L.C.}

\examples{
\donttest{RunStats()}
}

\keyword{misc}
