\name{Trends-package}

\alias{Trends-package}

\docType{package}

\title{
Analysis of Data Collected from a Monitoring Network
}

\description{
This package is for identifying trends in temporal observational data.
}

\details{
  \tabular{ll}{
    Package: \tab Trends\cr
    Type: \tab Package\cr
    Version: \tab 0.0-1\cr
    Date: \tab 2011-11-04\cr
    License: \tab GPL (>= 2)
  }
}

\note{
The set of standards used for coding \pkg{Trends} is documented in
\href{http://google-styleguide.googlecode.com/svn/trunk/google-r-style.html}{Google's R Style Guide}.
}

\author{
Jason C. Fisher

Maintainer: <\email{jfisher@usgs.gov}>
}

\keyword{package}

