\name{trends}
\alias{trends}
\alias{trends-package}
\alias{runApp}

\title{Trend Analysis Shiny App}

\description{
Shiny app to perform trend analyses for conformity and continuous data.
}

\usage{
runApp()
}

\author{Brecht Devleesschauwer (\email{brechtdv@gmail.com})}

\examples{
\dontrun{
runApp()
}
}
