\name{DrawPlot}

\alias{DrawPlot}

\title{Draw Single Plot}

\description{
This function draws a single plot on a graphics device.
}

\usage{
DrawPlot(d, tbl.par, xlim = c(NA, NA), ylim = c(NA, NA),
         main = NULL, xlab = NULL, ylab = NULL, leg.box.col="#FFFFFF")
}

\arguments{
\item{d}{data.frame; something.}
\item{tbl.par}{data.frame; something.}
\item{xlim}{POSIXct; something.}
\item{ylim}{numeric; something.}
\item{main}{character; something.}
\item{xlab}{character; something.}
\item{ylab}{character; something.}
\item{leg.box.col}{character; something.}
}

\details{
Something
}

\author{Fisher, J.C.}

\seealso{
\code{\link{plot.window}}, \code{\link{axis.POSIXct}}, \code{\link{legend}}
}

\examples{
print("something")
}

\keyword{hplot}
