% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawPlot.R
\name{DrawPlot}
\alias{DrawPlot}
\title{Draw Single Plot}
\usage{
DrawPlot(d, model, plim = c(0.1, 0.9), xlim = NULL, ylim = NULL,
  main = NULL, ylab = "")
}
\arguments{
\item{d}{data.frame.
Observations, where the first and second column is of class Date and \link{Surv}, respectively.
The Surv component is of type \emph{interval} and should not contain right-censored data.}

\item{model}{\link{survreg}.
Survival regression model}

\item{plim}{numeric.
Percentile limits defining the confidence band of the regression model.}

\item{xlim}{Date.
The \emph{x} limits of plot.}

\item{ylim}{numeric.
The \emph{y} limits of plot.}

\item{main}{character.
Main plot title}

\item{ylab}{character.
Label for \emph{y} axis.}
}
\value{
Used for the side-effect of a new plot generated.
}
\description{
This function draws a single time-series plot on the active graphics device.
}
\details{
Uncensored data is drawn as points.
  Left- and interval-censored data is drawn as vertical lines with short-horizontal lines at the interval bounds.
  Note that the lower bound of left-censored data is placed at zero.
  The 50th percentile of the regression model is drawn as a curved line.
  The confidence band is drawn as a solid polygon.
}
\examples{
# Create time-series data set with uncensored, left-censored, and interval-censored data
n <- 30
x <- as.Date(sort(sample(1:1000, n)), origin = as.Date("1992-02-15"))
time1 <- runif(n, min = 0, max = 100)
time2 <- time1 + runif(n, min = 0, max = 10)
time1[sample(1:n, 10)] <- NA
idxs <- sample(1:n, 10)
time1[idxs] <- time2[idxs]
y <- survival::Surv(time1, time2, type = "interval2")
d <- data.frame(x, y)

# Fit the data with a regression model
model <- survival::survreg(y ~ x, data = d)

# Plot the data and regression model
DrawPlot(d, model, plim = c(NA, 0.8), main = "Title", ylab = "Value")

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\seealso{
\code{\link{RunAnalysis}}, \code{\link[survival]{predict.survreg}}
}
\keyword{hplot}

