% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessConfig.R
\name{ProcessConfig}
\alias{ProcessConfig}
\title{Process Configuration}
\usage{
ProcessConfig(config, processed.obs)
}
\arguments{
\item{config}{data.frame.
See \sQuote{Details} section}

\item{processed.obs}{list.
See documentation for \code{\link{ProcessObs}} function for details.}
}
\value{
Returns a data.frame object with the following components:
  \describe{
    \item{Site_id}{unique site identifier}
    \item{Site_name}{local site name}
    \item{Parameter_id}{unique parameter identifier}
    \item{row}{row index in the \code{config} data table.}
  }
}
\description{
This function processes configuration records for data analysis.
}
\details{
Required columns in the \code{config} data table include:
  \code{Site_id}, a unique site identifier;
  \code{Site_name}, a local site name; and
  \code{Parameter_id}, unique parameter identifier(s).
  The \code{config} data table is composed of character-class components.
  Multiple parameter identifiers can be specified in a single value of
  \code{config$Parameter_id} using a comma separator.
  Site and parameter identifiers not found in the \code{processed.obs} list are removed.
}
\seealso{
\code{\link{RunAnalysis}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{methods}
