\name{RunTheilSen}

\alias{RunTheilSen}

\title{Run Theil-Sen Estimator}

\description{
This function computes  a .95 confidence interval for a Theil-Sen estimator.
}

\usage{
RunTheilSen(x, y, alpha = 0.5, pr = FALSE)
}

\arguments{
\item{x}{numeric; something.}
\item{y}{numeric; something.}
\item{alpha}{numeric; something.}
\item{pr}{logical; something.}
}

\details{
Something
}

\author{R.R. Wilcox and J.C. Fisher}

\examples{
print("something")
}

\keyword{misc}
