\name{ShowSymbols}

\alias{ShowSymbols}

\title{Open Graphics Device}

\description{
Draw all possible point symbols within a legend.
}

\usage{
ShowSymbols(file.symbs)
}

\arguments{
\item{file.symbs}{character; something.}
}

\details{
Something
}

\value{
Something
}

\author{Fisher, J.C.}

\seealso{
\code{\link{legend}}
}

\examples{
\donttest{ShowSymbols()}
}

\keyword{hplot}
